added config
