+++ src/main/java/org/zalando/nakadi/service/TracingService.java
@@ -1,12 +1,9 @@
 package org.zalando.nakadi.service;
 
 import com.google.common.collect.ImmutableMap;
-import com.lightstep.tracer.jre.JRETracer;
-import com.lightstep.tracer.shared.Options;
 import io.opentracing.Scope;
 import io.opentracing.Span;
 import io.opentracing.Tracer;
-import io.opentracing.noop.NoopTracerFactory;
 import io.opentracing.tag.Tags;
 import io.opentracing.util.GlobalTracer;
 import org.slf4j.Logger;
@@ -20,39 +17,17 @@ import javax.annotation.PostConstruct;
 @Configuration
 public class TracingService {
     private static final Logger LOG = LoggerFactory.getLogger(TracingService.class);
-    private String componentName = "mycomponent";
-    private String accessToken = "test";
-    private String collectorHost = "localhost";
-    private int collectorPort = 8444;
-    public final Tracer tracer;
+    private Tracer tracer;
 
     @Autowired
-    public TracingService() {
-        this.tracer = tracer();
+    public TracingService(final Tracer tracer) {
+        this.tracer = tracer;
     }
 
     @PostConstruct
     public void registerGlobalTracer() {
         if (!GlobalTracer.isRegistered()) {
             GlobalTracer.register(tracer);
-        }
-    }
-
-    public Tracer tracer() {
-        try {
-            final Options options = new Options.OptionsBuilder()
-                    .withAccessToken(accessToken)
-                    .withCollectorHost(collectorHost)
-                    .withCollectorPort(collectorPort)
-                    .withComponentName(componentName)
-                    .build();
-
-            final Tracer tracer = new JRETracer(options);
-            LOG.info("Initialized Lightstep Tracer");
-            return tracer;
-        } catch (Exception ex) {
-            LOG.error("Invalid Lightstep configuration. Returning a NoopTracer. {}", ex.getMessage());
-            return NoopTracerFactory.create();
         }
     }
