+++ src/main/resources/application.yml
@@ -37,6 +37,11 @@ nakadi:
     dataType: service
     value: stups_nakadi
   authz.warnAllDataAccessMessage: "Data access warning"
+  tracing:
+    componentName: "nakadi"
+    accessToken: ""
+    collectorHost: "localhost"
+    collectorPort: 8444
   topic:
     min:
       retentionMs: 10800000 # 3 hours
